Core: Refactor a lot of timer work and fix a couple of bugs EE/IOP Timers: improve clock sync, disable v/h sync when SINT enabled.
Some changes based on tests from PS2

[SAVEVERSION+]

+++ pcsx2/CDVD/CDVD.h
@@ -123,7 +123,7 @@ struct cdvdStruct
 
 	// Calculates the number of Vsyncs and once it reaches a total number of Vsyncs worth a second with respect to
 	// the videomode's vertical frequency, it updates the real time clock.
-	int RTCcount;
+	double RTCcount;
 	cdvdRTC RTC;
 
 	u32 CurrentSector;
